+++ plugins/dates/src/api/parse/one/01-tokenize/03-time.js
@@ -66,13 +66,12 @@ const parseTime = function (doc, context) {
   time = time.not('^(at|by|for|before|this|after)')
   time = time.not('sharp')
   time = time.not('on the dot')
-  // time.debug()
+
   let s = spacetime.now(context.timezone)
   let now = s.clone()
   // check for known-times (like 'today')
-  let timeStr = time.not('in the').text('reduced')
+  let timeStr = time.not('in? the').text('reduced')
   timeStr = timeStr.replace(/^@/, '')//@4pm
-  // console.log(timeStr)
   if (hardCoded.hasOwnProperty(timeStr)) {
     return { result: hardCoded[timeStr], m: time }
   }